Use reference counting instead of Strings
I'm not sure if this is better or not. It's slightly more complicated
but it seems to line up better against the [anti-patterns](https://yew.rs/docs/next/concepts/function-components/properties#anti-patterns)
section of the docs. This program is small enough that this won't really
matter, but I'd like to do things the right way. I'm going to leave this
branch hanging here until I decide what makes more sense.