Round start_datetime and end_datetime to nearest minute client-side
The html input only allows input rounded to the nearest minute, but when
we submitted the inital value of Local::now() it included seconds and
miliseconds that we usually couldn't have the specificity for. We still
allow any valid datetime as the server because if a client says that's
when it started then that's when it started. But we'd like the data from
our own client to always be consistent.